Deprecate macro `#method` in favour of `#through`
As explained in #1 by @dgmstuart, the `#method` in spryte has a nameclash with `Object#method`.

Chose to name it to `#through` rather than `#http_method` or `#verb` since it's a semantically intuitive DSL.

Made sure that we output a deprecation warning pointing users to the new method and that the test suite will fail once we upgrade major version without removing the deprecated method.